Two big changes.
First, there is a new exported ConfigFile struct that contains all
user-configurable data. This is so that you can create projects that
basically consist of a config file and import this project and set
the config; then you have a branded project. It's more limited than
you might initially think, but it adds little cost.

Second, the template handlers now emit errors to StdErr upon
failure, so gone are the days when a templating failure would leave
you with a blank webpage and a blank console.